requirements.txt
torch
torchvision
streamlit
streamlit-drawable-canvas
numpy
pillow